# Projetos2024/2M/cozinhaembyte/public/api/openapi.yaml
openapi: "3.0.1"
info:
  title: Receitas Culinarias
  description: >
    Aplicação de Receitas Culinárias.
    Um website de receitas culinárias é uma aplicação em que usuários divulgam suas receitas para outros usuários da rede.
    Esses receitas possuem título, descrição, ingredientes, modo de preparo, tempo de preparo e uma imagem correspondente.
    As receitas também podem ser avaliadas pelos usuários, com estrelas de 1 a 5, o que aparece é a média das avaliações.
    As receitas podem ser comentadas por outros usuários, para construir uma feedback da receita ao dono da receita.
    Nessa versão, há cadastro de todas as entidades com suporte geral, mas não há segurança de dados.
  version: "2.0"
servers:
  - url: http://{host}:{porta}/api
    variables:
      host:
        default: localhost
      porta:
        default: "8000"
paths:
...
  /isAuthenticated:
    post:
      operationId: isAuthenticated
      summary: Verificar se o usuário está autenticado
      security:
        - BearerAuth: []
      responses:
        200:
          description: Verificação realizada
          content:
            application/json:
              schema:
                type: boolean
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT